backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
pyjwt==2.10.1
bcrypt==4.2.1
python-multipart==0.0.20
httpx==0.28.1
paramiko==3.5.1
cryptography==44.0.0
pyyaml==6.0.2
aiofiles==24.1.0
redis[hiredis]==5.2.1
anthropic==0.40.0
openai==1.58.1
google-generativeai==0.8.4
websockets==14.1
